+++ packages/superset-ui-demo/storybook/stories/plugins/preset-chart-xy/BoxPlot/stories/Basic.tsx
