+++ packages/superset-ui-demo/storybook/stories/plugins/preset-chart-xy/Line/Stories.tsx
@@ -13,4 +13,3 @@ export { default as withLabelFlush } from './stories/flush';
 export { default as withMissingData } from './stories/missing';
 export { default as legacyShim } from './stories/legacy';
 export { default as withTimeShift } from './stories/timeShift';
-
