+++ packages/superset-ui-demo/storybook/stories/plugins/preset-chart-xy/Line/stories/missing.tsx
@@ -1,31 +1,34 @@
-import * as React from 'react';
+import React from 'react';
 import { SuperChart } from '@superset-ui/chart';
-import { radios } from '@storybook/addon-knobs';
 import data from '../data/data';
 import { LINE_PLUGIN_TYPE } from '../constants';
 import dummyDatasource from '../../../../../shared/dummyDatasource';
 
-export default () => (
+const missingData = data.map(({ y, ...rest }) => ({
+  ...rest,
+  y: Math.random() < 0.25 ? null : y,
+}));
+
+const missing = () => (
   <SuperChart
     key="line1"
     chartType={LINE_PLUGIN_TYPE}
+    width={400}
+    height={400}
     datasource={dummyDatasource}
+    queryData={{ data: missingData }}
     formData={{
       encoding: {
         x: {
           field: 'x',
           type: 'temporal',
-          format: '%Y-%m',
+          format: '%Y',
           scale: {
             type: 'time',
           },
           axis: {
-            orient: radios('x.axis.orient', { top: 'top', bottom: 'bottom' }, 'bottom'),
-            title: radios(
-              'x.axis.title',
-              { enable: 'Time', disable: '', '': undefined },
-              'Time',
-            ),
+            orient: 'bottom',
+            title: 'Time',
           },
         },
         y: {
@@ -35,27 +38,19 @@ export default () => (
             type: 'linear',
           },
           axis: {
-            orient: radios(
-              'y.axis.orient',
-              { left: 'left', right: 'right', '': undefined },
-              'left',
-            ),
-            title: radios(
-              'y.axis.title',
-              { enable: 'Score', disable: '', '': undefined },
-              'Score',
-            ),
+            orient: 'left',
+            title: 'Score',
           },
         },
         stroke: {
           field: 'name',
           type: 'nominal',
+          scale: {},
           legend: true,
         },
       },
     }}
-    height={400}
-    queryData={{ data }}
-    width={400}
   />
 );
+
+export default missing;
