+++ packages/superset-ui-demo/storybook/stories/plugins/preset-chart-xy/Line/stories/timeShift.tsx
@@ -1,22 +1,17 @@
 import React from 'react';
 import { SuperChart } from '@superset-ui/chart';
-import data from '../data/data';
+import data from '../data/data2';
 import { LINE_PLUGIN_TYPE } from '../constants';
 import dummyDatasource from '../../../../../shared/dummyDatasource';
 
-const missingData = data.map(({ y, ...rest }) => ({
-  ...rest,
-  y: Math.random() < 0.25 ? null : y,
-}));
-
-const missing = () => (
+export default () => (
   <SuperChart
     key="line1"
     chartType={LINE_PLUGIN_TYPE}
     width={400}
     height={400}
     datasource={dummyDatasource}
-    queryData={{ data: missingData }}
+    queryData={{ data }}
     formData={{
       encoding: {
         x: {
@@ -43,14 +38,41 @@ const missing = () => (
           },
         },
         stroke: {
-          field: 'name',
+          value: '#1abc9c',
           type: 'nominal',
-          scale: {},
-          legend: true,
+          scale: false,
+        },
+        fill: {
+          field: 'snapshot',
+          type: 'nominal',
+          scale: {
+            type: 'ordinal',
+            domain: ['Current', 'Last year'],
+            range: [true, false],
+          },
+          legend: false,
+        },
+        strokeDasharray: {
+          field: 'snapshot',
+          type: 'nominal',
+          scale: {
+            type: 'ordinal',
+            domain: ['Current', 'Last year'],
+            range: [null, '4 4'],
+          },
+          legend: false,
+        },
+        strokeWidth: {
+          field: 'snapshot',
+          type: 'nominal',
+          scale: {
+            type: 'ordinal',
+            domain: ['Current', 'Last year'],
+            range: [3, 1.5],
+          },
+          legend: false,
         },
       },
     }}
-  />,
+  />
 );
-
-export default missing;
