+++ packages/superset-ui-demo/storybook/stories/plugins/preset-chart-xy/ScatterPlot/stories/bubble.tsx
@@ -23,7 +23,7 @@ export default () => (
             type: 'linear',
           },
           axis: {
-            orient: radios('x.axis.orient', ['top', 'bottom'], 'bottom'),
+            orient: radios('x.axis.orient', { top: 'top', bottom: 'bottom' }, 'bottom'),
           },
         },
         y: {
@@ -33,7 +33,7 @@ export default () => (
             type: 'linear',
           },
           axis: {
-            orient: radios('y.axis.orient', ['left', 'right'], 'left'),
+            orient: radios('y.axis.orient', { left: 'left', right: 'right' }, 'left'),
           },
         },
         size: {
